refactor: 패키지 분리 - MVC 기반
- domain
- controller
- view
