+++ README.md
@@ -59,8 +59,8 @@
 
 ## step5 - 자동차 경주(리펙토링)
 ###  리팩토링 요구사항
-- [ ] domain 패키지 분리
-- [ ] view 패키지 분리
+- [x] domain 패키지 분리
+- [x] view 패키지 분리
 - [ ] MVC 기반으로 리팩토링 하기, 
   - view 패키지의 객체는 domain 에 의존가능
   - domain 패키지의 객체는 view 에 의존 불가능 가능
